+++ src/SparkPost/AsyncRequestSender.cs
@@ -1,0 +1,36 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace SparkPost
+{
+    public class AsyncRequestSender : IRequestSender
+    {
+        private readonly Client client;
+
+        public AsyncRequestSender(Client client)
+        {
+            this.client = client;
+        }
+
+        public async Task<Response> Send(Request request)
+        {
+            using (var httpClient = client.CustomSettings.CreateANewHttpClient())
+            {
+                httpClient.BaseAddress = new Uri(client.ApiHost);
+                httpClient.DefaultRequestHeaders.Accept.Clear();
+                httpClient.DefaultRequestHeaders.Add("Authorization", client.ApiKey);
+
+                var result = await new RequestMethodFinder(httpClient)
+                    .FindFor(request)
+                    .Execute(request);
+
+                return new Response
+                {
+                    StatusCode = result.StatusCode,
+                    ReasonPhrase = result.ReasonPhrase,
+                    Content = await result.Content.ReadAsStringAsync()
+                };
+            }
+        }
+    }
+}
