+++ src/SparkPost/RequestSender.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 
 namespace SparkPost
 {
@@ -8,34 +7,10 @@ namespace SparkPost
         Task<Response> Send(Request request);
     }
 
-    public class RequestSender : IRequestSender
+    public class RequestSender : AsyncRequestSender
     {
-        private readonly Client client;
-
-        public RequestSender(Client client)
+        public RequestSender(Client client) : base(client)
         {
-            this.client = client;
-        }
-
-        public async Task<Response> Send(Request request)
-        {
-            using (var httpClient = client.CustomSettings.CreateANewHttpClient())
-            {
-                httpClient.BaseAddress = new Uri(client.ApiHost);
-                httpClient.DefaultRequestHeaders.Accept.Clear();
-                httpClient.DefaultRequestHeaders.Add("Authorization", client.ApiKey);
-
-                var result = await new RequestMethodFinder(httpClient)
-                    .FindFor(request)
-                    .Execute(request);
-
-                return new Response
-                {
-                    StatusCode = result.StatusCode,
-                    ReasonPhrase = result.ReasonPhrase,
-                    Content = await result.Content.ReadAsStringAsync()
-                };
-            }
         }
     }
 }
